Move general class to utils
